Further refined the discount implementation to parameterise discount processor for Item by weight or Item by Unit.

+++ src/main/java/kata/supermarket/discount/DiscountProcessor.java
@@ -2,6 +2,7 @@ package kata.supermarket.discount;
 
 import kata.supermarket.Item;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 /**
@@ -9,6 +10,5 @@ import java.util.List;
  */
 public interface DiscountProcessor {
 
-    int calculateDiscount(List<Item> discounts);
-
+    BigDecimal calculateDiscount(List<Item> discounts);
 }
